[Decomp][Spell] Cohesion split: extract SpellAuraMisc handlers to SpellAuraMisc.cpp
